working service to keep app alive
