+++ kotlin/KeepAliveService.kt
@@ -1,0 +1,54 @@
+package dev.dioxus.main
+
+import dev.dioxus.main.RustWebView
+import dev.dioxus.main.Logger
+import android.annotation.SuppressLint
+import android.os.Build
+import android.os.Bundle
+import android.webkit.WebView
+import android.view.KeyEvent
+import androidx.appcompat.app.AppCompatActivity
+import android.app.Notification
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.Service
+import android.content.Intent
+import android.os.IBinder
+import android.util.Log
+
+class KeepAliveService : Service() {
+    override fun onCreate() {
+        super.onCreate()
+        startForegroundWithNotification()
+    }
+
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        return START_STICKY
+    }
+
+    override fun onBind(intent: Intent?): IBinder? = null
+
+    private fun startForegroundWithNotification() {
+        val channelId = "media_channel"
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val channel = NotificationChannel(
+                channelId,
+                "Rust Service",
+                NotificationManager.IMPORTANCE_LOW
+            )
+            val manager = getSystemService(NotificationManager::class.java)
+            manager.createNotificationChannel(channel)
+        }
+
+        val notification: Notification = Notification.Builder(this, channelId)
+            .setContentTitle("Rust Service Running")
+            .setContentText("Your Rust code is active")
+            .setSmallIcon(android.R.drawable.ic_media_play)
+            .build()
+
+        startForeground(1, notification)
+
+        Log.i("com.example.Music", "hi")
+    }
+}
