+++ kotlin/MainActivity.kt
@@ -19,162 +19,5 @@ import android.util.Log
 import com.example.Trackfish.BuildConfig;
 typealias BuildConfig = BuildConfig;
 
-class MainActivity : MusicActivity()
+class MainActivity : WryActivity() 
 
-abstract class MusicActivity : AppCompatActivity() {
-    private lateinit var mWebView: RustWebView
-
-    open fun onWebViewCreate(webView: WebView) { }
-
-    fun setWebView(webView: RustWebView) {
-        mWebView = webView
-        onWebViewCreate(webView)
-    }
-
-    val version: String
-        @SuppressLint("WebViewApiAvailability", "ObsoleteSdkInt")
-        get() {
-            // Check getCurrentWebViewPackage() directly if above Android 8
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                return WebView.getCurrentWebViewPackage()?.versionName ?: ""
-            }
-
-            // Otherwise manually check WebView versions
-            var webViewPackage = "com.google.android.webview"
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-              webViewPackage = "com.android.chrome"
-            }
-            try {
-                @Suppress("DEPRECATION")
-                val info = packageManager.getPackageInfo(webViewPackage, 0)
-                return info.versionName.toString()
-            } catch (ex: Exception) {
-                Logger.warn("Unable to get package info for '$webViewPackage'$ex")
-            }
-
-            try {
-                @Suppress("DEPRECATION")
-                val info = packageManager.getPackageInfo("com.android.webview", 0)
-                return info.versionName.toString()
-            } catch (ex: Exception) {
-                Logger.warn("Unable to get package info for 'com.android.webview'$ex")
-            }
-
-            // Could not detect any webview, return empty string
-            return ""
-        }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        create(this)
-
-        Log.i("com.example.Music", "woah 0")
-        val intent = Intent(this, KeepAliveService::class.java)
-        Log.i("com.example.Music", "woah 1")
-        startForegroundService(intent)
-        Log.i("com.example.Music", "woah 2")
-    }
-
-    override fun onStart() {
-        super.onStart()
-        start()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        resume()
-    }
-
-    override fun onPause() {
-        super.onPause()
-        pause()
-    }
-
-    override fun onStop() {
-        super.onStop()
-        stop()
-    }
-
-    override fun onWindowFocusChanged(hasFocus: Boolean) {
-        super.onWindowFocusChanged(hasFocus)
-        focus(hasFocus)
-    }
-
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
-        save()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        destroy()
-        onActivityDestroy()
-    }
-
-    override fun onLowMemory() {
-        super.onLowMemory()
-        memory()
-    }
-
-    override fun onKeyDown(keyCode: Int, event: KeyEvent?): Boolean {
-        if (keyCode == KeyEvent.KEYCODE_BACK && mWebView.canGoBack()) {
-            mWebView.goBack()
-            return true
-        }
-        return super.onKeyDown(keyCode, event)
-    }
-
-    fun getAppClass(name: String): Class<*> {
-        return Class.forName(name)
-    }
-
-    companion object {
-        init {
-            System.loadLibrary("dioxusmain")
-        }
-    }
-
-    private external fun create(activity: MusicActivity)
-    private external fun start()
-    private external fun resume()
-    private external fun pause()
-    private external fun stop()
-    private external fun save()
-    private external fun destroy()
-    private external fun onActivityDestroy()
-    private external fun memory()
-    private external fun focus(focus: Boolean)
-}
-
-class KeepAliveService : Service() {
-    override fun onCreate() {
-        super.onCreate()
-        createNotificationChannel()
-
-        val notification: Notification = Notification.Builder(this, "media_channel")
-            .setContentTitle("Trackfish is running")
-            .setContentText("Your music service is active")
-            .setSmallIcon(android.R.drawable.ic_media_play)
-            .build()
-
-        startForeground(1, notification)
-
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        return START_STICKY
-    }
-
-    override fun onBind(intent: Intent?): IBinder? = null
-
-    private fun createNotificationChannel() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(
-                "media_channel",
-                "Trackfish Service",
-                NotificationManager.IMPORTANCE_LOW
-            )
-            getSystemService(NotificationManager::class.java)?.createNotificationChannel(channel)
-        }
-    }
-}
