+++ kotlin/WryActivity.kt
@@ -7,12 +7,20 @@
 package dev.dioxus.main
 
 import dev.dioxus.main.RustWebView
+import dev.dioxus.main.Logger
 import android.annotation.SuppressLint
 import android.os.Build
 import android.os.Bundle
 import android.webkit.WebView
 import android.view.KeyEvent
 import androidx.appcompat.app.AppCompatActivity
+import android.app.Notification
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.Service
+import android.content.Intent
+import android.os.IBinder
+import android.util.Log
 
 abstract class WryActivity : AppCompatActivity() {
     private lateinit var mWebView: RustWebView
@@ -59,7 +67,16 @@ abstract class WryActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        create(this)
+        create(this)        
+        Intent().setClassName("com.example.Trackfish", "dev.dioxus.main.KeepAliveService")
+        val intent = Intent(this, KeepAliveService::class.java)
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            startForegroundService(intent)
+        } else {
+            startService(intent)
+        }
+        Log.i("com.example.Music", "started KeepAliveService")
     }
 
     override fun onStart() {
@@ -114,13 +131,6 @@ abstract class WryActivity : AppCompatActivity() {
     fun getAppClass(name: String): Class<*> {
         return Class.forName(name)
     }
-
-    companion object {
-        init {
-            System.loadLibrary("dioxusmain")
-        }
-    }
-
     private external fun create(activity: WryActivity)
     private external fun start()
     private external fun resume()
@@ -132,5 +142,10 @@ abstract class WryActivity : AppCompatActivity() {
     private external fun memory()
     private external fun focus(focus: Boolean)
 
-    
+    companion object {
+        init {
+            System.loadLibrary("dioxusmain")
+        }
+    }
 }
+
